Add files via upload
Added example.fs and example.txt.  Running example.fs will result in output similar to what is in example.txt.
Creates, and displays,  two linked lists of numbers.  Calculates, and displays, a list of the intersection.  
Calculates, and displays, a list of the union.  Displays the memory stores usage.  Deallocates the four lists.
Displays the memory stores usage (should be zero in use).