Add files via upload
In the function .mma-usage, I changed "Free", the current number of free array items, to "Min Free", the minimum number of free items during the run of the program.  So you can see how close the program came to stack underflow.  The old "Free" is obvious as "Capacity" - "In use".